refactor(pages): folder structure
