+++ src/App.js
@@ -1,10 +1,10 @@
 import { Fragment } from 'react';
 import { Routes, Route, Navigate } from 'react-router-dom';
 
-import Home from './pages/Home';
-import Portfolio from './pages/Portfolio';
-import About from './pages/About';
-import Contact from './pages/Contact';
+import Home from './pages/Home/Home';
+import Portfolio from './pages/Portfolio/Portfolio';
+import About from './pages/About/About';
+import Contact from './pages/Contact/Contact';
 
 import Navbar from './components/Navbar/Navbar';
 import Footer from './components/Footer/Footer';
